Move websocket handler to separate file
